+++ android/app/src/main/java/com/project/libum/core/exeption/IncorrectPasswordException.kt
@@ -1,0 +1,3 @@
+package com.project.libum.core.exeption
+
+class IncorrectPasswordException(message: String) : Exception(message)
