+++ android/app/src/main/java/com/project/libum/di/AppModule.kt
@@ -2,9 +2,8 @@ package com.project.libum.di
 
 import android.app.Application
 import com.project.libum.BuildConfig
-import com.project.libum.data.remote.ApiService
-import com.project.libum.domain.recaptcha.ExecuteRecaptchaUseCase
-import com.project.libum.domain.recaptcha.InitializeRecaptchaUseCase
+import com.project.libum.domain.usecase.ExecuteRecaptchaUseCase
+import com.project.libum.domain.usecase.InitializeRecaptchaUseCase
 import com.project.libum.domain.recaptcha.RecaptchaService
 import com.project.libum.domain.recaptcha.RecaptchaServiceImpl
 import com.project.libum.domain.repository.AuthRepository
@@ -12,7 +11,6 @@ import com.project.libum.domain.repository.AuthRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityComponent
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
