+++ android/app/src/main/java/com/project/libum/domain/repository/AuthRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.project.libum.domain.repository
 
+import com.project.libum.core.exeption.IncorrectPasswordException
 import com.project.libum.data.model.LoginRequest
 import com.project.libum.data.model.LoginResponse
 import com.project.libum.data.remote.ApiClient.apiService
@@ -13,7 +14,11 @@ class AuthRepositoryImpl: AuthRepository {
             if (response.isSuccessful) {
                 Result.success(response.body()!!)
             } else {
-                Result.failure(Exception("Login failed"))
+                if(response.code() == 401){
+                    Result.failure(IncorrectPasswordException("Incorrect password"))
+                }else{
+                    Result.failure(Exception("Login failed"))
+                }
             }
         } catch (e: Exception) {
             Result.failure(e)
