+++ android/app/src/main/java/com/project/libum/domain/usecase/ExecuteRecaptchaUseCase.kt
@@ -1,6 +1,7 @@
-package com.project.libum.domain.recaptcha
+package com.project.libum.domain.usecase
 
 import com.google.android.recaptcha.RecaptchaAction
+import com.project.libum.domain.recaptcha.RecaptchaService
 
 class ExecuteRecaptchaUseCase(private val recaptchaService: RecaptchaService) {
     suspend operator fun invoke(action: RecaptchaAction): Result<Unit> {
