+++ android/app/src/main/res/values-night/themes.xml
@@ -1,7 +1,6 @@
 <resources>
     <!-- Base application theme. -->
     <style name="Theme.Libum" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
-
         <item name="colorPrimary">@color/blue_500</item>
         <item name="colorPrimaryVariant">@color/blue_700</item>
         <item name="colorOnPrimary">@color/white</item>
