+++ android/app/src/main/res/values/colors.xml
@@ -16,5 +16,10 @@
     <color name="gray_blue_200">#A5C6FA</color>
     <color name="gray_blue_500">#819CC7</color>
     <color name="gray_blue_700">#576883</color>
+
+    -- Additional colors --
+
+    <color name="blue_100">#DDE4F0</color>
+
 </resources>
